requirements.txt
requests
beautifulsoup4
lxml
trafilatura
markdownify
pyyaml
sqlalchemy
python-dateutil
openai
python-dotenv
